Add plugin uninstall; install exact versions
of plugins. The previous installation would post to
pluginManager/installNecessaryPlugins, but this endpoint will always
ignore the requested version and install the latest version. Previous
versions can only be installed by uploading the hpi/jpi file for that
version. To get the previous versions, the current version number in the
url will be replaced by the requested version number.
The plugin will be downloaded, unpacked and eventual dependencies (at
their latest version) will be installed.
Intoduceed ability to uninstall plugins and option to wait for reboot
when calling safe_restart().